修改开始时间字段类型

+++ app/src/main/java/com/hdcy/app/model/Bean4VedioBanner.java
@@ -21,7 +21,7 @@ public class Bean4VedioBanner implements Serializable {
     public int viewCount;
     public boolean live;
     public String start;
-    public int startTime;
+    public long startTime;
     public String length;
     public int commentCount;
